+ Moved database connector to a separate file. + Added dependencies to README.
