+++ README.md
@@ -1,2 +1,6 @@
 # Initial tests for NL2SQL processing
 Author: Frank William Hammond Espinosa
+
+## Dependencies:
++ python-dotenv
++ mysql.connector (for now)
